requirements.txt
fastapi
uvicorn[standard]
python-dotenv
scikit-learn
joblib
pandas
numpy
requests
